\docType{data}
\name{data_122}
\alias{data_122}
\title{Simulated egame122 data}
\usage{
data(data_122)
}
\description{
Simulated data for illustrating \code{\link{egame122}}.
}
\details{
The variables are: \describe{
\item{\code{f1}, \code{f2}}{Factors with levels "a", "b", "c"}
\item{\code{x1}--\code{x5}}{Numeric variables entering Player 1's utilities}
\item{\code{z1}--\code{z3}}{Numeric variables entering Player 2's utilities}
\item{\code{a1}}{Indicator for Player 1's move (L or R)}
\item{\code{a2}}{Indicator for Player 2's move (L or R)}
\item{\code{y}}{Factor containing outcome}
}
}
\seealso{
\code{\link{egame122}}
}
\keyword{data}

