\docType{data}
\name{data_123}
\alias{data_123}
\title{Simulated egame123 data}
\usage{
data(data_123)
}
\description{
Simulated data for illustrating \code{\link{egame123}}.
}
\details{
The variables are: \describe{
\item{\code{x1}--\code{x8}}{Regressors}
\item{\code{a1}}{Indicator for Player 1's move (L or R)}
\item{\code{a2}}{Indicator for Player 2's move (L or R)}
\item{\code{a3}}{Indicator for Player 3's move (L or R)}
\item{\code{y}}{Numeric variable containing outcome number: 1, 3, 5, or 6,
corresponding to labels in the game tree in the \code{\link{egame123}}
documentation.}
}
}
\seealso{
\code{\link{egame123}}
}
\keyword{data}

