\docType{data}
\name{data_ult}
\alias{data_ult}
\title{Simulated ultimatum data}
\usage{
data(data_ult)
}
\description{
Simulated data for illustrating \code{\link{ultimatum}}.
}
\details{
The variables are: \describe{
\item{\code{offer}}{The offer made by Player 1}
\item{\code{accept}}{Whether Player 2 accepted the offer (0 for rejection, 1
for acceptance)}
\item{\code{w1}, \code{w2}}{Variables entering both players' reservation values}
\item{\code{x1}--\code{x4}}{Variables entering Player 1's reservation value}
\item{\code{z1}--\code{z4}}{Variables entering Player 2's reservation value}
}
The maximum offer size is 15.
}
\seealso{
\code{\link{ultimatum}}
}
\keyword{data}

