\name{print.game}
\alias{print.game}
\title{Print a strategic model object}
\usage{\method{print}{game}(x, ...)
}
\description{The default method for printing a \code{game} object.}
\details{Prints the call and coefficients of a fitted strategic model.}
\value{\code{x}, invisibly}
\author{Brenton Kenkel (\email{brenton.kenkel@gmail.com})}
\arguments{\item{x}{a fitted model of class \code{game}}
\item{...}{other arguments, currently ignored}
}

