\name{print.summary.game}
\alias{print.summary.game}
\title{Print strategic model summary}
\usage{\method{print}{summary.game}(x, ...)
}
\description{Print output from \code{summary.game}}
\details{Prints the standard regression results table from a fitted strategic model,
along with the log-likelihood, AIC, and number of observations.}
\value{\code{x}, invisibly.}
\seealso{\code{\link{summary.game}}}
\author{Brenton Kenkel (\email{brenton.kenkel@gmail.com})}
\arguments{\item{x}{an object of class \code{summary.game}, typically produced by
\code{summary} on a fitted model of class \code{game}}
\item{...}{other arguments, currently ignored}
}

