\docType{data}
\name{student_offers}
\alias{student_offers}
\title{Data from students playing the ultimatum game}
\usage{
data(student_offers)
}
\description{
Data from a trial of the ultimatum game with graduate students.
}
\details{
The variables are: \describe{
\item{\code{offer}}{The offer made by Player 1}
\item{\code{accept}}{Whether Player 2 accepted the offer (0 for rejection, 1
for acceptance)}
\item{\code{gender1}}{Whether Player 1 is female}
\item{\code{gender2}}{Whether Player 2 is female}
}
The maximum offer size is 100.
}
\seealso{
\code{\link{ultimatum}}
}
\keyword{data}

