% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/games.r
\docType{package}
\name{games-package}
\alias{games-package}
\title{A package for estimating strategic statistical models.}
\description{
A package for estimating strategic statistical models.
}
\section{Acknowledgements}{
 We thank the Wallis Institute of Political
Economy for financial support.
}
\references{
Brenton Kenkel and Curtis S. Signorino.  2014.  "Estimating Extensive Form
Games in R."  \emph{Journal of Statistical Software} 56(8):1--27.

Curtis S. Signorino.  2003.  "Structure and Uncertainty
in Discrete Choice Models."  \emph{Political Analysis} 11:316--344.
}

