\docType{package}
\name{games-package}
\alias{games-package}
\title{A package for estimating strategic statistical models.}
\description{
  A package for estimating strategic statistical models.
}
\section{Acknowledgements}{
  We thank the Wallis Institute of Political Economy for
  financial support.
}
\references{
  Curtis S. Signorino.  2003.  "Structure and Uncertainty
  in Discrete Choice Models."  \emph{Political Analysis}
  11:316--344.
}

