\name{print.game}
\alias{print.game}
\title{Print a strategic model object}
\usage{
\method{print}{game}(x, ...)
}
\arguments{
  \item{x}{a fitted model of class \code{game}}

  \item{...}{other arguments, currently ignored}
}
\value{
\code{x}, invisibly
}
\description{
The default method for printing a \code{game} object.
}
\details{
Prints the call and coefficients of a fitted strategic model.
}
\author{
Brenton Kenkel (\email{brenton.kenkel@gmail.com})
}

